+++ lib/core/snackbar_notifier.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 
-import '../screens/global_components/snackbar.dart';
+import '../components/components.dart';
 import 'core.dart';
 
 class SnackBarNotifier extends ConsumerWidget {
