+++ lib/databse_service.dart
@@ -1,10 +1,8 @@
 import 'package:isar/isar.dart';
 import 'package:path_provider/path_provider.dart';
-import 'package:sickler/models/models.dart';
-import 'package:sickler/models/water/water_log.dart';
-import 'package:sickler/models/water/water_preferences.dart';
 
 import 'features/auth/models/user/app_user.dart';
+import 'features/water/water.dart';
 
 class LocalDbService {
   ///Pass all collection Schemas here
@@ -13,7 +11,7 @@ class LocalDbService {
 
     if (Isar.instanceNames.isEmpty) {
       return await Isar.open(
-          [SicklerUserSchema, WaterPreferencesSchema, WaterLogSchema],
+          [AppUserSchema, WaterPreferencesSchema, WaterLogSchema],
           inspector: true, directory: directory.path);
     }
     return Future.value(Isar.getInstance());
