+++ lib/features/auth/repositories/user/user_repository.dart
@@ -3,12 +3,10 @@ import 'dart:developer';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:sickler/core/core.dart';
-import 'package:sickler/services/user/remote/user_service.dart';
-
 import '../../auth.dart';
-import '../../models/models.dart';
 import '../../services/auth/remote/auth_service.dart';
 import '../../services/user/local/user_local_service.dart';
+import '../../services/user/remote/user_service.dart';
 
 class UserRepository {
   final UserService _userService;
