+++ lib/features/auth/screens/auth/auth_success.dart
@@ -5,6 +5,8 @@ import 'package:gap/gap.dart';
 import 'package:go_router/go_router.dart';
 import 'package:sickler/core/constants.dart';
 
+import '../../../../components/components.dart';
+import '../../../profile/profile.dart';
 import '../../auth.dart';
