+++ lib/features/auth/screens/auth/register_screen.dart
@@ -5,7 +5,9 @@ import 'package:gap/gap.dart';
 import 'package:go_router/go_router.dart';
 import 'package:sickler/core/core.dart';
 
+import '../../../../components/bottom_nav_bar.dart';
 import '../../../../components/components.dart';
+import '../../../profile/profile.dart';
 import '../../auth.dart';
