+++ lib/features/auth/services/user/remote/user_service.dart
@@ -1,6 +1,6 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 
-import '../../../models/models.dart';
+import '../../../auth.dart';
 
 class UserService {
   final FirebaseFirestore firestore = FirebaseFirestore.instance;
