+++ lib/features/emergency/screens/components/crisis_history_monthly.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:gap/gap.dart';
-import 'package:sickler/screens/global_components/components.dart';
+
+import '../../../../components/components.dart';
 
 class CrisisHistoryMonthly extends StatelessWidget {
   const CrisisHistoryMonthly({super.key});
