+++ lib/features/emergency/screens/components/crisis_history_weekly.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:gap/gap.dart';
-import 'package:sickler/screens/global_components/components.dart';
 
+import '../../../../components/components.dart';
 import 'components.dart';
 
 class CrisisHistoryWeekly extends StatelessWidget {
