+++ lib/features/emergency/screens/components/crisis_log_card.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_animate/flutter_animate.dart';
 import 'package:gap/gap.dart';
 
-import '../../../core/core.dart';
+import '../../../../core/core.dart';
 
 class CrisisLogCard extends StatefulWidget {
   const CrisisLogCard({super.key});
