+++ lib/features/emergency/screens/components/dialogs/emergency_alert_select_contact_dialog.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:gap/gap.dart';
 import 'package:sickler/core/core.dart';
-import 'package:sickler/screens/global_components/components.dart';
 
+import '../../../../../components/components.dart';
 import '../components.dart';
 
 class EmergencyAlertSelectContactDialog extends StatelessWidget {
