+++ lib/features/emergency/screens/components/dialogs/emergency_select_contacts_dialog.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:gap/gap.dart';
 import 'package:sickler/core/core.dart';
-import 'package:sickler/screens/global_components/components.dart';
+
+import '../../../../../components/components.dart';
 
 class EmergencySelectContactsDialog extends StatelessWidget {
   const EmergencySelectContactsDialog({super.key});
