+++ lib/features/emergency/screens/emergency_screen.dart
@@ -2,10 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:gap/gap.dart';
 import 'package:go_router/go_router.dart';
 import 'package:sickler/core/core.dart';
-import 'package:sickler/screens/emergency/add_emergency_contact_screen.dart';
-import 'package:sickler/screens/emergency/components/components.dart';
 
-import '../global_components/components.dart';
+import '../../../components/components.dart';
+import '../emergency.dart';
 
 class EmergencyScreen extends StatelessWidget {
   static const String id = "emergency";
