+++ lib/features/home/screens/components/emergency_sharing_card.dart
@@ -3,7 +3,6 @@ import 'package:flutter_svg/flutter_svg.dart';
 import 'package:gap/gap.dart';
 
 import '../../../../core/core.dart';
-import '../../../core/core.dart';
 
 class EmergencySharingCard extends StatelessWidget {
   final VoidCallback onPressed;
