+++ lib/features/home/screens/components/feeling_card.dart
@@ -1,7 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:gap/gap.dart';
 
-import '../../../core/core.dart';
+import '../../../../core/core.dart';
+
 
 class FeelingCard extends StatefulWidget {
   const FeelingCard({super.key});
