+++ lib/features/home/screens/components/meds_reminder_card.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
-import '../../../core/core.dart';
+import '../../../../core/core.dart';
 import 'components.dart';
 
 class MedsReminderCard extends StatelessWidget {
