+++ lib/features/home/screens/components/water_card.dart
@@ -2,9 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:gap/gap.dart';
 
+import '../../../../components/components.dart';
 import '../../../../core/core.dart';
-import '../../../core/core.dart';
-import '../../global_components/components.dart';
 
 class WaterCard extends StatelessWidget {
   final VoidCallback onPressed;
