+++ lib/features/water/services/remote/water_service.dart
@@ -1,8 +1,5 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:intl/intl.dart';
-import 'package:sickler/models/water/water_preferences.dart';
-
-import '../../../models/water/water_log.dart';
 import '../../water.dart';
 
 class WaterService {
